reorganised code to use one console client, put patterns in thier categories
