--- clients/AdapterClient/Program.cs
@@ -1,16 +1,0 @@
-﻿using ClassAdapter;
-using static System.Console;
-
-Title = "Adapter Pattern";
-
-ICityAdapter adapter   =  new CityAdapter();
-var city = adapter.GetCity();
-
-WriteLine($"{city.FullName}, {city.Inhabitants}");
-
-ReadKey();
-
-
-
-
-
