+++ clients/ConsoleClient/Examples/BuilderDemo.cs
@@ -1,0 +1,34 @@
+﻿using static System.Console;
+using Builder;
+
+namespace ConsoleClient.Examples
+{
+    static class BuilderDemo
+    {
+        public static void Run()
+        {
+          Title = "Builder";
+
+            // The garage is our director. It is where we build our cars
+            var garage = new Garage();
+
+            // we are building two brands
+            var miniBuilder = new MiniBuilder();
+            var bmwBuilder = new BMWBuilder();
+
+            garage.Construct(miniBuilder);
+            garage.Show();
+
+
+
+            garage.Construct(bmwBuilder);
+            garage.Show();
+
+            //Console.WriteLine(miniBuilder.Car.ToString());
+            //Console.WriteLine(bmwBuilder.Car.ToString());
+
+
+            Console.ReadKey();
+        }
+    }
+}
