+++ clients/ConsoleClient/Program.cs
@@ -2,6 +2,10 @@
 
 //Test Patterns
 
+FactoryMethodDemo.Run();
+FacadeDemo.Run();
+AdapterDemo.Run();
+BuilderDemo.Run();
 AbstractFactoryDemo.Run();
 
 //BridgeDemo.Run();
