+++ src/Structural/Adapter/Adapter.csproj
