+++ src/Structural/Adapter/ClassAdapterImplementation.cs
