+++ src/Structural/Adapter/ObjectAdapterImplementation.cs
