+++ src/Structural/Facade/Implementation.cs
